fix(ci): resolve compile errors caught by Actions
- label the onGestureDispatched lambda explicitly (property assignment
  has no implicit return label)
- add missing Spacer/height imports in GazeCalibrationScreen
- coerce Double p95 error to Float for savedP95

+++ app/src/main/java/com/aircontrol/ui/gazecalibration/GazeCalibrationViewModel.kt
@@ -305,7 +305,7 @@ class GazeCalibrationViewModel @Inject constructor(
         if (fit.isSuccess && fit.model != null) {
             settingsRepository.updatePersonalizedGazeCalibration(fit.model.toSerialized())
             anySaved = true
-            savedP95 = fit.model.validationMetrics.p95NormalizedError
+            savedP95 = fit.model.validationMetrics.p95NormalizedError.toFloat()
             Timber.i(
                 "Personalized gaze calibration saved (%d samples, val p95=%.4f)",
                 fit.model.validationMetrics.sampleCount,
